refactor: 상아 피드백 반영해서 컴포넌트 depth 1로 고정 (#230)
