--- src/pages/OnboardingPage/Step/index.ts
@@ -1,3 +1,0 @@
-export { default as StepStart } from './StepStart/StepStart.tsx';
-export { default as StepField } from './StepField/StepField.tsx';
-export { default as StepService } from './StepService/StepService.tsx';
