+++ src/pages/OnboardingPage/StepService/BoxAllowedService/BoxAllowedService.tsx
